fix(ci): split test_meter_provider.py to satisfy pylint max-module-lines
Root cause: the module reached 1018 lines, above pylint's
max-module-lines=1000 (C0302), causing the lint job (and the alls-green
check gate) to fail. Move the self-contained TestExemplarFilter class
into its own test file with only the imports it needs; no test logic
changed.
